% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishMO2-package.R
\docType{data}
\name{codSDA}
\alias{codSDA}
\title{Oxygen uptake data from a juvenile Atlantic cod before and after ingestion
of a meal}
\format{
A data frame with 947 observations on the following 5 variables:
\describe{ \item{DateTime}{a POSIXct vector indicating date and
time for each measurement} \item{Logtime_hr}{a numeric vector
indicating the time, in hours, relative to feeding. Measurements obtained
before feeding have a negative value for this variable}
\item{MO2}{a numeric vector representing oxygen uptake, in
micromoles of oxygen per min per kg of body mass}
\item{MO2cor}{same as MO2, except for the first 10 hours after the
meal: these measurements were corrected for the increase in MO2 resulting
from the feeding procedure. The correction was based on an exponential
decay function fitted to data obtained after a sham feeding that took place
the day prior to feeding.} \item{r2}{a numeric vector representing
the coefficient of determination R^2 of the regression between dissolved
oxygen and time for a given value of MO2} }
}
\source{
Collected by Denis Chabot at the Maurice-Lamontagne Institute,
Department of Fisheries and Oceans, Canada.
}
\description{
Oxygen uptake data from a juvenile (22.9 g) Atlantic cod (\emph{Gadus
morhua}) measured by intermittend-flow respirometry before and after
ingestion of a meal (about 3.5\% body mass, experimental diet). This is to
illustrate \verb{calcSDA} and \verb{plotSDA}.
}
\details{
Intermittent flow respirometry and a 2.56 L respirometer were used to
measure oxygen uptake in this juvenile cod (salinity of 27⋅6 ± 1⋅7, 11⋅8 ±
0⋅1∘ C, mean ± s.d.). The fish was fasted for 66 h before being placed into
the respirometer to insure it was in post-absorptive state, a requirement
to measure its standard metabolic rate. After 3 days, it was taken out of
the respirometer, anasthetised, a tube was inserted in its stomach and
water was flushed in its stomach (sham feeding). Approximately 24 h later,
the same procedure was repeated but water was replaced by a paste made of
the experimental pellets and water. Approximately 3.5\% of body mass was
fed to the fish, which was then returned to the respirometers. Oxygen
uptake was measured for an additional 4 days.

This fish was placed into the respirometer on the 2009-01-05 and removed on
2009-01-13. The first MO2 after feeding was measured 16:01:12 on
2009-01-09, which was taken at time zero. The first and last MO2 measures
were obtained at -97.22 and 88.84 h. Sham-feeding was completed at -25.65
h. Examination of the raw data suggested that only MO2 measurements with a
$r^2$ > 0.96 should be considered valid. This particular fish acclimated
slowly to the respirometer and it is recommended to exclude the first 24-h
of data before calculating SMR. Data recorded during the first 10-h after
sham-feeding or after feeding should also be excluded when calculating SMR.
}
\examples{

data(codSDA)
plot(MO2~DateTime, data=codSDA)

}
\references{
Chabot, Denis, Koenker, Roger and Farrell, Anthony P. (2016)
The measurement of the specific dynamic action in fishes.  \emph{Journal of
Fish Biology} 88, 152-172.
}
\keyword{datasets}
