% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convO2conc.R
\name{convO2conc}
\alias{convO2conc}
\title{Conversion between oxygen concentration units.}
\usage{
convO2conc(
  x,
  from = c("mg", "ml", "umol", "mmol"),
  to = c("umol", "mg", "ml", "mmol")
)
}
\arguments{
\item{x}{A single value or a vector of values to be converted.}

\item{from}{Oxygen concentration unit of x. Must be one of "mg", "ml",
"umol" [for micromol] or "mmol".}

\item{to}{Desired oxygen concentration unit. Must be one of "mg", "ml",
"umol" [for micromol] or "mmol".}
}
\value{
Returns a vector of converted values with same length as the vector
x.
}
\description{
Converts between different conversion units (mg, ml, micromol, mmol)
}
\details{
Conversion from one unit of oxygen concentration to another.
}
\examples{

convO2conc(10)
convO2conc(10, from="mg", to="umol")
convO2conc(250, from="umol", to="mg")

}
\references{
García, H. E. and Gordon, L. I. (1992) Oxygen solubility in
seawater: better fitting equations.  \emph{Limnology and Oceanography} 37,
1307-1312.
}
\author{
Denis Chabot, Maurice-Lamontagne Institute, DFO, Canada.
}
