% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishMO2-package.R
\docType{package}
\name{fishMO2-package}
\alias{fishMO2-package}
\alias{fishMO2}
\title{Calculate and plot the standard metabolic rate (SMR), the critical oxygen
level (O2crit) and the specific dynamic action (SDA) and related variables
in fishes and crustaceans}
\description{
Package to calculate and plot the metabolic rate of water-breathers such as
fish and crustaceans. In particular, functions are provided to calculate
the Standard Metabolic Rate (SMR) from multiple measurements of oxygen
uptake obtained from a single animal. In experiments where ambient
dissolved oxygen was slowly decreased until the animal could no longer
sustain its SMR, O2crit can be calculated. Specific Dynamic Action (SDA)
can also be calculated and plotted for experiments measuring oxygen uptake
before and after a meal. Conversion between different oxygen concentration
units is provided, functions to calculate oxygen solubility and saturation
values will be added soon.
}
\references{
Chabot, D., Steffensen, J. F. and Farrell, A. P. (2016) The
determination of the standard metabolic rate in fishes.  \emph{Journal of
Fish Biology} 88, 81-121.

Claireaux, Guy and Chabot, Denis (2016) Responses by fishes to
environmental hypoxia: integration through Fry's concept of aerobic
metabolic scope.  \emph{Journal of Fish Biology} 88, 232-251.
doi:doi:10.1111/jfb.1283
}
\author{
Denis Chabot

Maintainer: Denis Chabot \email{chabot.denis@gmail.com}
}
\keyword{package}
