% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishMO2-package.R
\docType{data}
\name{GrHalO2crit}
\alias{GrHalO2crit}
\title{Oxygen uptake data from an adult Greenland halibut to illustrate the
determination of SMR and O2crit}
\format{A data frame with 220 observations on the following 6 variables:
\describe{ \item{list("DateTime")}{a POSIXct vector indicating date and
time for each measurement} \item{list("Logtime_hr")}{a numeric vector
indicating the time, in hours, relative to feeding. Measurements obtained
before feeding have a negative value for this variable}
\item{list("MO2")}{a numeric vector representing oxygen uptake, in
micromoles of oxygen per min per kg of body mass} \item{list("r2")}{a
numeric vector representing the coefficient of determination R^2 of the
regression between dissolved oxygen and time for a given value of MO2}
\item{list("Aver_O2")}{a numeric vector representing the average level of
dissolved oxygen (in kPa) in the respirometer during each measurement of
MO2} \item{list("DO")}{a numeric vector representing the equivalent of
Aver_O2 in \% saturation units} }}
\source{
Collected by Denis Chabot at the Maurice-Lamontagne Institute,
Department of Fisheries and Oceans, Canada.
}
\description{
Oxygen uptake data from an adult (1.563 kg) Greenland halibut
(\emph{Reinhardtius hippoglossoides}) measured by intermittend-flow
respirometry. During the last few hours of the experiment, the ambient
dissolved oxygen was slowly reduced until the fish was no longer able to
sustain its Standard Metabolic Rate (SMR), so as to calculate its critical
oxygen level (O2crit).
}
\details{
Results of an experiment with a Greenland halibut \emph{Reinhardtius
hippoglossoides} of 1.563 kg placed in a 48.43 L respirometer. Oxygen
uptake was measured by intermittent-flow respirometry for about 2.5 days.
These measurements allowed the determination of the Standard Metabolic Rate
(SMR) by the method described in Chabot et al. 2016, using the
\verb{calcSMR} function from this package. At this point, dissolved oxygen
was decreased slowly until the fish could no longer sustain its SMR. The
\verb{calcO2crit} function from this package can be used to determine
O2crit (in \% saturation).
}
\examples{

data(GrHalO2crit)
plot(MO2~DateTime, data=GrHalO2crit)

}
\references{
Chabot, D., Steffensen, J. F. and Farrell, A. P. (2016) The
determination of the standard metabolic rate in fishes.  \emph{Journal of
Fish Biology} 88, 81-121.

Claireaux, Guy and Chabot, Denis (2016) Responses by fishes to
environmental hypoxia: integration through Fry's concept of aerobic
metabolic scope.  \emph{Journal of Fish Biology} 88, 232-251.
doi:doi:10.1111/jfb.1283
}
\keyword{datasets}
