% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeO2lab.R
\name{makeO2lab}
\alias{makeO2lab}
\title{Make a formatted label for oxygen uptake (MO2), SDA or oxygen
concentration.}
\usage{
makeO2lab(mo2 = c("dotital", "dot", "ital", "plain", "hidden"),
  o2 = c("mg", "ug", "mmol", "umol", "ml", "mL"), t = c("hr", "min"),
  m = c("kg", "g", "mg"), v = c("l", "L"), showO2 = F,
  Type = c("MO2msp", "MO2ind", "SDAmsp", "SDAind", "conc"))
}
\arguments{
\item{mo2}{If the label is for oxygen uptake (see Type), selects the
formatting of the expression MO2, such as \verb{dotital} (adds a dot on
italics M to express a rate), \verb{dot}, \verb{ital} or \verb{plain}. The
expression MO2 can also by \verb{hidden}. Not used for SDA or concentration
labels.}

\item{o2}{Oxygen units, either \verb{mg}, \verb{ug} (gives microg),
\verb{mmol}, \verb{umol} (gives micromol), \verb{ml} or \verb{mL}.}

\item{t}{Time unit, either \verb{hr} or \verb{min}.}

\item{m}{Mass unit for the animal, either \verb{kg}, \verb{g} or \verb{mg}.
Used for mass-specific MO2 or SDA, not used in "per individual"
measurements.}

\item{showO2}{Logical, to hide or show \verb{O2} after the oxygen unit.}

\item{Type}{Select the type of label, either for mass-specific MO2
(\verb{MO2msp}), individual MO2 (\verb{MO2ind}), mass-specific SDA
magnitude (\verb{SDAmsp}), individual SDA magnitude (\verb{SDAind}) or
oxygen concentration (\verb{conc}).}
}
\value{
Returns the formatted label.
}
\description{
Plot labels involving oxygen usually require subscript, superscript, italic
or dot formatting. This function facilitates the creation of such labels.
}
\details{
The combination of R language and Tex commands required to obtain complex
labels can be difficult to master (and remember). This single function
allows the user to format a wide variety of labels dealing with MO2, SDA,
oxygen concentration to produce publication-quality figures.
}
\examples{

# for a plot of MO2 as a function of time
mo2lab = makeO2lab(mo2="dotital", o2= "mg", t="hr", m="kg", v="l", showO2 = F, Type="MO2msp")
plot(1:10, xlab="Time", ylab=mo2lab, mgp=c(2,0.5,0))

}
\author{
Denis Chabot, Institut Maurice-Lamontagne, Fisheries and Oceans
Canada.
}
