% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTZs.R
\docType{data}
\name{myTZs}
\alias{myTZs}
\title{List of time zones.}
\format{An object of class \code{list} of length 8.}
\usage{
myTZs
}
\value{
A list with 8 elements:

\item{IMLcivil}{IML (Institut Maurice-Lamontagne, Maurice Lamontagne
Institute), uses daylight saving time in Summer.} \item{IMLst}{IML
(Institut Maurice-Lamontagne, Maurice Lamontagne Institute), uses standard
time year long.} \item{FRcivil}{France, civil time, i.e. uses daylight
saving time in Summer.} \item{FRst}{France, standard time, i.e. does not
use daylight saving time in Summer.} \item{DKcivil}{Danemark, civil time,
i.e. uses daylight saving time in Summer.} \item{DKst}{Danemark, standard
time, i.e. does not use daylight saving time in Summer.}
\item{BCcivil}{British Columbia, civil time, i.e. uses daylight saving time
in Summer.} \item{BCst}{British Columbia, standard time, i.e. does not use
daylight saving time in Summer.}
}
\description{
List of time zones at sites where respirometry data were available to test
this package.
}
\details{
Can be used to specify time zone in \code{drawNights}. The choice of using
daylight saving time or not must match how the computer was setup when
respirometry data were obtained.
}
\examples{

myTZs$IMLst

}
\keyword{datasets}
