% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.R
\name{N}
\alias{N}
\title{Calculates the number of non-missing values}
\usage{
N(x)
}
\arguments{
\item{x}{A vector of numeric or character values that may include NAs.}
}
\value{
An integer representing the number of non-missing values in
\verb{x}.
}
\description{
Calculates the number of non-missing values in the elements of vector
\verb{x}.
}
\details{
Used by \verb{calcSMR}.
}
\examples{

a = c(1:5, NA, 7:10)
N(a)

}
