% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotO2crit.R
\name{plotO2crit}
\alias{plotO2crit}
\title{Plot oxygen uptake values as a function of dissolved oxygen and show
O2crit.}
\usage{
plotO2crit(
  o2critobj,
  plotID = "",
  o2lab = "Dissolved oxygen (\% sat.)",
  mo2lab = expression(M[O[2]]),
  smr.cex = 0.9,
  o2crit.cex = 0.9,
  plotID.cex = 1.2,
  Transparency = T,
  ...
)
}
\arguments{
\item{o2critobj}{Object created by \verb{calcO2crit}. It is a list that
contains the original data, in addition to the SMR and O2crit.}

\item{plotID}{Can be used to add a label (e.g., A, B, C, a, b, c) to the
upper-left corner of a plot. For a finer control, produce the plot without
plotID and add it manually afterward.}

\item{o2lab}{Character string to be used to label the X-axis.}

\item{mo2lab}{Character string to be used to label the Y-axis. The user can
create an object with \verb{makeO2lab} and use it here for a more
sophisticated result than the default.}

\item{smr.cex}{The size (cex) of the SMR value, shown on the plot.}

\item{o2crit.cex}{The size (cex) of the O2crit value, shown on the plot.}

\item{plotID.cex}{The size (cex) of the plot label, if shown on the plot.}

\item{Transparency}{Some of the colours are transparent, which is not
supported by all graphics devices. For \verb{postscript} and other such
devices, set \verb{Transparency} to FALSE and similar but not transparent
colours will be used instead.}

\item{\dots}{Additional graphics parameters accepted by \verb{plot}.}
}
\value{
Produces a plot.
}
\description{
Takes an object created by \verb{calcO2crit}. This object is self-contained
and includes the original data, SMR and O2crit. This functions produces a
plot showing the original values of oxygen uptake as a function of
dissolved oxygen, along with an horizontal line representing SMR and a
regression line fitted to those oxygen uptake values deemed to be
conforming (related to dissolved oxygen). The O2crit is shown as the
intersection of the two lines.
}
\details{
Produces a plot illustrating how O2crit was calculated for a given animal,
based on data obtained by intermittent-flow respirometry. Typically the
data set includes multiple measurements of oxygen uptake in conditions that
are compatible with the determination of the Standard Metabolic Rate (SMR).
Then the ambient level of dissolved oxygen is slowly decreased until the
animal can no longer sustain its SMR. The only data required are contained
in an object created by \verb{calcO2crit}. Depending on the number of
values of oxygen uptake identified as conforming (proportional) to
dissolved oxygen level and below SMR, on the necessity (or not) to include
a value of oxygen uptake far above SMR to increase the slope and on the
type of regression fitted to these values the color and type of regression
line changes. For ordinary LS regression, without added points: blue line,
red symbols; ordinary LS regression, with an added point: blue line, red
symbols, except orange for the added value; regression through origin,
green dotted line, red symbols.
}
\note{
%% ~~further notes~~
}
\section{Warning }{
 For R plotting devices that do not support transparency,
set Transparency to FALSE.
}

\examples{

data("GrHalO2crit")
# remove a 10-h acclimation period; also remove the last 11 h, when DO was decreasing to calculate O2crit. There were no low R^2s that required removal
smrData = subset(GrHalO2crit, DateTime >= DateTime[1] + 10*60*60 & DateTime <= DateTime[nrow(GrHalO2crit)] - 11*60*60)
smr = calcSMR(smrData$MO2)  # contains several estimations of SMR, need to chose one
SMR = as.numeric(ifelse(smr$CVmlnd > 5.4, smr$quant[4], smr$mlnd)); SMR # as recommended in Chabot et al. 2016
MyO2crit = calcO2crit(GrHalO2crit, SMR)
plotO2crit(MyO2crit)
plotO2crit(MyO2crit, plotID="A")

}
\references{
Chabot, Denis, Steffensen, John F. and Farrell, Anthony P.
(2016) The determination of the standard metabolic rate in fishes.
\emph{Journal of Fish Biology} 88, 81-121.

Claireaux, Guy and Chabot, Denis (2016) Responses by fishes to
environmental hypoxia: integration through Fry's concept of aerobic
metabolic scope.  \emph{Journal of Fish Biology} 88, 232-251.
doi:doi:10.1111/jfb.1283
}
\seealso{
\code{\link{calcO2crit}}, \code{\link{calcSMR}}
}
\author{
Denis Chabot, Institut Maurice-Lamontagne, Department of Fisheries
and Oceans.
}
