% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSDA.R
\name{plotSDA}
\alias{plotSDA}
\title{Add SDA fit and SDA parameters to an existing plot of oxygen uptake values
as a function of time.}
\usage{
plotSDA(my.sda, show.par = T, show.var = T, show.smr.line = T,
  show.feeding = T, show.mag.arrow = T, peak.lab.pos = NULL,
  mag.lab.pos = NULL, smr.lab.pos = c("2L", "1L", "none"),
  lty.smr = 1, col.smr = "orange", lty.feeding = 1,
  Col.feeding = "seagreen", Cex.par = 0.8, Cex.sda = 0.9,
  Col.sda = rgb(1, 0, 0, 0.2), ...)
}
\arguments{
\item{my.sda}{Object created by \code{calcSDA}.}

\item{show.par}{Boolean, if \code{TRUE}, displays in upper-left of plot the
values of tau and lambda that were used to obtain the fitted line.}

\item{show.var}{Boolean, if \code{TRUE}, displays the values of peak MO2
(net, i.e. after removal of SMR), time to peak, duration and magnitude of
SDA.}

\item{show.smr.line}{Boolean, \code{TRUE} displays a horizontal line at SMR.}

\item{show.feeding}{Boolean, \code{TRUE} displays a vertical line at time
zero to illustrate feeding time.}

\item{show.mag.arrow}{Boolean, \code{TRUE} displays an arrow from the word
"magnitude" to the shaded area representing the area under the curve.}

\item{peak.lab.pos}{If NULL, the function tries to calculate a position for
the label associated with the peak of SDA. A (x, y) pair can be provided
if the calculated position is inadequate.}

\item{mag.lab.pos}{If NULL, the function tries to calculate a position for
the label associated with the magnitude of SDA. A (x, y) pair can be
provided if the calculated position is inadequate.}

\item{smr.lab.pos}{Although a horizontal line depicting SMR is shown, the
value of SMR is only shown if requested, either in one string below SMR
(\code{1L}) or in two parts, one above and one below SMR (\code{2L}).}

\item{lty.smr}{Line type for SMR, same meaning as \code{lty} for any line in
R.}

\item{col.smr}{Colour of the horizontal line representing SMR.}

\item{lty.feeding}{Line type for time of feeding, same meaning as \code{lty}
for any line in R.}

\item{Cex.par}{Size of the text for tau and lambda. Same use as \code{cex}
parameter in \code{text}.}

\item{Cex.sda}{Size of the text for the SDA parameters. Same use as
\code{cex} parameter in \code{text}}

\item{Col.sda}{Colour of the shading representing SDA magnitude.}

\item{\dots}{Other graphics parameters may work.}

\item{col.feeding}{Colour of the vertical line representing feeding time.}
}
\description{
Displays the SDA line fitted by \code{calcSDA} on top of a plot of MO2 vs
time. Also displayed is coloured representation of the magnitude of SDA (the
area between the fitted line and the horizontal line representing SMR) and
the values of peak, time-to-peak, duration and magnitude.
}
\details{
Add SDA details to an existing plot of MO2 x time.
}
\note{
%% ~~further notes~~
}
\section{Warning}{
 The automatic positioning of SMR and of the SDA parameters
  works best when plots have a longer X axis than Y axis. On square plots
  you will probably have to position these labels yourself using the
  arguments \code{peak.lab.pos} and \code{mag.lab.pos}. On plots showing
  data before and after feeding, the space to draw an arrow associated with
  magnitude (\code{show.mag.arrow}) is restricted and the result is often
  unattractive, it may be necessary to set this argument to \code{FALSE}.
}

\examples{

data("codSDA")
sda.data = subset(codSDA, r2 >= 0.96)
# Logtime_hr is time in hours relative to time of feeding, a required variable.
# MO2cor is in micromoles per min per kg and the "cor" in the name means that
#    the MO2 values soon after feeding were corrected for the effect of the
#    feeding procedure

SDA= calcSDA(X=sda.data$Logtime_hr, Y=sda.data$MO2cor, my.smr=81.2,
               tau=0.2, lambda=30,
	           postfeed.acclim=3, tol=0.05, tol.type="proportion",
	           MO2.time.unit="min", X.time.unit="hour")

par(mar=c(4, 4, 0.1, 0.1)+0.1, mgp=c(1.8,0.5,0))
plotMO2(sda.data$Logtime_hr, sda.data$MO2cor, mo2="dotital", o2="umol", t="min", m="kg",
    showO2 = F, Xlab = "Time relative to feeding (hr)", ylim=c(50, 250))
plotSDA(SDA)
# not enough space to show the arrow associated with magnitude, better to hide it
plotMO2(sda.data$Logtime_hr, sda.data$MO2cor, mo2="dotital", o2="umol", t="min", m="kg",
    showO2 = F, Xlab = "Time relative to feeding (hr)", ylim=c(50, 250))
plotSDA(SDA, show.mag.arrow=F)

# same data, but focus on post-feeding data and the SDA
plotMO2(sda.data$Logtime_hr, sda.data$MO2cor, mo2="dotital", o2="umol", t="min", m="kg",
    showO2 = F, Xlab = "Time relative to feeding (hr)", ylim=c(50, 250), xlim=c(0,100))
plotSDA(SDA, smr.lab.pos="1L")

# SMR label is illegible, better not to display it
# Default position for peak not great, force a new position
plotMO2(sda.data$Logtime_hr, sda.data$MO2cor, mo2="dotital", o2="umol", t="min", m="kg",
    showO2 = F, Xlab = "Time relative to feeding (hr)", ylim=c(50, 250), xlim=c(0,100))
plotSDA(SDA, smr.lab.pos="none", peak.lab.pos=c(40,150))



}
\references{
Chabot, Denis, Koenker, Roger and Farrell, Anthony P. (2016) The
  measurement of the specific dynamic action in fishes.  \emph{Journal of
  Fish Biology} 88, 152-172.

  Chabot, Denis, Steffensen, John F. and Farrell, Anthony P. (2016) The
  determination of the standard metabolic rate in fishes.  \emph{Journal of
  Fish Biology} 88, 81-121.

  Koenker, Roger (2005) Quantile regression.  Cambridge University Press,
  Cambridge. 366 p.
}
\seealso{
\code{\link{calcSDA}}, \code{\link{plotMO2}}, \code{\link{text}}
}
\author{
Denis Chabot, Institut Maurice-Lamontagne, Department of Fisheries
  and Oceans.
}
\keyword{~kwd1}
\keyword{~kwd2}
