% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po2.R
\name{po2}
\alias{po2}
\title{Pressure of dissolved oxygen in air or water}
\usage{
po2(t, s = 0, p = 101.325)
}
\arguments{
\item{t}{A vector of temperature values in degrees Celsius}

\item{s}{A vector of salinity values (parts per thousand)}

\item{p}{Total gas pressure (in kPa), typically atmospheric pressure above
the body of water}
}
\value{
Partial pressure of oxygen in kPa
}
\description{
Calculate oxygen partial pressure as a function of temperature and
salinity. Oxygen is a constant fraction (o2frac) of total gas pressure
after removing water vapor pressure.
}
\details{
See Warnings.
}
\note{
%% ~~further notes~~
}
\section{Warnings}{
 It is assumed that vectors \code{t} and \code{s} have
the same length, or that one of the two is a constant value. This is not
checked in the present version, the user must make sure it is the case.
}

\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as

po2(t=10, s=28, p=101.325)


}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Denis Chabot
}
\keyword{~kwd1}
\keyword{~kwd2}
