% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sites.R
\docType{data}
\name{Sites}
\alias{Sites}
\title{List of lat-long for known sites.}
\format{
An object of class \code{list} of length 6.
}
\usage{
Sites
}
\value{
A list with 6 elements (i.e. locations). For each element, lat and
long (in decimal degrees) are provided as a vector with 2 elements:

\item{La_Rochelle}{La Rochelle, France.} \item{Sete}{Sete, France.}
\item{IML}{IML (Institut Maurice-Lamontagne, Maurice Lamontagne Institute),
Fisheries and Oceans, QC, Canada.} \item{Helsingor}{Helsingør, Danemark.}
\item{Brest}{Brest, France.} \item{Vancouver}{Vancouver, BC, Canada.}
}
\description{
List containing geographic position (latitude and longitude, in decimal
degrees) of sites where respirometry data were available to test this
package.
}
\details{
Can be used to specify location in \code{drawNights}. Latitude and
longitude are in decimal degrees.
}
\examples{

Sites$IML

}
\keyword{datasets}
