% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solO2.R
\name{solO2}
\alias{solO2}
\title{Solubility of oxygen in fresh- and seawater}
\usage{
solO2(
  t,
  s = 0,
  o2Unit = c("mg", "ml", "umol"),
  waterUnit = c("l", "kg"),
  p = 101.325,
  author = c("Garcia", "Benson")
)
}
\arguments{
\item{t}{A vector of temperature values in degrees Celsius}

\item{s}{A vector of salinity values}

\item{o2Unit}{The desired oxygen units for solubility calculation (mg, ml,
µmol)}

\item{waterUnit}{Unit of water used for solubility calculation (liters or
kg)}

\item{p}{Total gas pressure above the water body (kPa)}

\item{author}{Equations from 2 authors can be used, Garcia or Benson (see
references)}
}
\value{
Solubility value(s) of oxygen for requested temperatures and
salinities, in desired oxygen and water units, adjusted to pressure
requested by user.
}
\description{
Calculates the amount of oxygen that dissolves in water according to
temperature, salinity and atmospheric pressure, when water and air are in
equilibrium (100\% saturation). Different oxygen units are allowed.
Solubility can be calculated per liter (common in physiology) or per kg
(common in oceanography) of water.
}
\details{
Different units of oxygen concentration are allowed, different units for
quantity of water (1 liter or one kg), and equations from different two
authors are allowed. If total gas pressure is not 101.325, solubility will
be calculated for the pressure requested by the user.

The equations provided by García and Gordon (1992) are used by default,
they are supposedly the best available for seawater below 5 °C. Equations
of Benson and Krause (1980, 1984) are provided for comparison with the
results from DOTABLES program at http://water.usgs.gov/software/DOTABLES/.
For T varying from 0 to 15°C, P from 99 to 102 kPa, S=0 and 28 and values
rounded to 2 decimals, author set to "Benson", results match what DOTABLES
does.
}
\note{
%% ~~further notes~~
}
\section{Warnings}{
 It is assumed that vectors \code{t} and \code{s} have
the same length, or that one of the two is a constant value. This is not
checked in the present version, the user must make sure it is the case.
}

\examples{

solO2(t=5, s=28, o2Unit="mg")
# should give 10.5928 mg O2 per liter

# to compute \% saturation

obsDO = 8 # observed DO in mg per liter
Sol = solO2(t=10, s=28, o2Unit="mg")
sat = obsDO / Sol * 100
sat

}
\references{
Benson B. B. and Krause D. Jr. (1980) The concentration and
isotopic fractionation of gases dissolved in freshwater in equilibrium with
the atmosphere. 1. Oxygen.  \emph{Limnology and Oceanography}, 25, 662-671.

Benson B. B. and Krause D. Jr. (1984) The concentration and isotopic
fractionation of oxygen dissolved in freshwater and seawater in equilibrium
with the atmosphere.  \emph{Limnology and Oceanography}, 29, 620-632.

García, H. E. and Gordon, L.I. (1992) Oxygen solubility in seawater: better
fitting equations.  \emph{Limnology and Oceanography} 37, 1307-1312.
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Denis Chabot
}
\keyword{~kwd1}
\keyword{~kwd2}
