% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wvp.R
\name{wvp}
\alias{wvp}
\title{Water vapor pressure as a function of water temperature and salinity}
\usage{
wvp(t, s = 0, method = c("IAPWS", "BensonKrause", "GreenCarritt",
  "WMO2008", "Antoine", "Weiss", "GoffGratch"), sal_method = c("Chabot",
  "Green", "none"))
}
\arguments{
\item{t}{Temperature in °C}

\item{s}{Salinity in parts per thousand}

\item{method}{One of "IAPWS", "BensonKrause", "GreenCarritt","WMO2008",
"Antoine", "Weiss", "GoffGratch". See Details.}

\item{sal_method}{One of "Chabot", "Green" or "none". See Details.}
}
\value{
A vector of same length as parameter t
}
\description{
Calculates water vapor pressure, given water temperature and salinity,
using a variety of methods. This is required to convert partial pressure of
oxygen into saturation or concentration units and vice-versa.
}
\details{
Output: water vapor pressure above fresh or sea water in kPa.

Note: this function is not intended to calculate vapor pressure above ice.
Also, water vapor pressure is not affected by atmospheric pressure

Methods for vapor pressure calculation above fresh water:

\describe{ \item{IAPWS}{ This is the 'official' formulation from the
International Association for the Properties of Water and Steam (Wagner &
Pruss 1993), Eq. 2.5 p. 399. The valid range of this formulation is 273.16
<= K <= 647.096 and is based on the ITS90 temperature scale. }

\item{BensonKrause}{ Eq. 23 of Benson and Krause, 1980 }

\item{GreenCarritt}{ Green & Carritt 1967 (see also table 2 Benson & Krause
1984) without salinity correction, in atm converted to kPa }

\item{WMO2008}{ Annex 4B, Guide to Meteorological Instruments and Methods
of Observation, WMO Publication No 8, 7th edition, Geneva, 2008. }

\item{Antoine}{ http://www.watervaporpressure.com }

\item{Weiss}{ Weiss & Price 1980 eq 10 p. 350, in atm converted to kPa }

\item{GoffGratch}{ Smithsonian Meteorological Tables, 5th edition, p. 350,
1984, in hPa then converted to kPa } }

Methods for salinity correction: \describe{ \item{Chabot}{ My own
polynomial fit passing through origin to Robinson 1954's data, Table II p.
451 gives values of delta vp for different chlorinities. Chlorinity changed
into salinity according to Lyman (1969) }

\item{Green}{ From table 2 in Benson & Krause (1984) }

\item{none}{ Removes salinity correction, same as leaving \code{s} to its
default value of zero } }
}
\note{
This assumes that air temperature above the water surface is that of
the water. The effect of salinity is negligible, but it is nevertheless
included in the calculation if the user provides salinity.
}
\section{Warnings}{
 \code{s} must be of length 1 or same length as t. This
is not checked by the function at present so errors could occur if this is
not respected.
}

\examples{

wvp(t=c(0,5,10))
wvp(t=c(0,5,10), s=35)

}
\references{
Benson B.B. and Krause D.Jr. (1980) The concentration and
isotopic fractionation of gases dissolved in freshwater in equilibrium with
the atmosphere. 1. Oxygen.  \emph{Limnology and Oceanography}, 25, 662-671.

Benson B.B. and Krause D.Jr. (1984) The concentration and isotopic
fractionation of oxygen dissolved in freshwater and seawater in equilibrium
with the atmosphere.  \emph{Limnology and Oceanography}, 29, 620-632.

Green, E. J. and Carritt, D. E. (1967) New tables for oxygen saturation of
seawater.  \emph{Journal of Marine Research}, 25, 140-147.

Lyman, J. (1969) Redefinition of salinity and chlorinity.  \emph{Limnology
and Oceanography}, 27, 928-929.

Robinson R. A. (1954) The vapour pressure and osmotic equivalence of sea
water.  \emph{Journal of the Marine Biological Association of the United
Kingdom}, 33, 449-455.

Wagner, W. and Pruss, A. (1993) International Equations for the Saturation
Properties of Ordinary Water Substance.  Revised According to the
International Temperature Scale of 1990. Addendum to J. Phys. Chem. Ref.
Data 16, 893 (1987) \emph{Journal of Physical and Chemical Reference Data},
22, 783.  http://dx.doi.org/10.1063/1.555926.

Wagner, W. & Pruss, A. (2002) The IAPWS formulation 1995 for the
thermodynamic properties of ordinary water substance for general and
scientific use.  \emph{Journal of Physical and Chemical Reference Data},
31(2), 387-535.
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Denis Chabot
}
